separate rdist into a separate package
